refactor: group AACI/AAPB/OASW/sp-attack into views/utils/combat with shared predicates
Following the aaci.ts split, gather all battle-mechanics logic into a
single views/utils/combat package so the four modules share one
predicate pool instead of duplicating it:

- types.ts: merged GameShip/GameEquip object types (hoisted from aaci)
- combinators.ts: generic builders (shipIdIs, ctypeIs, itemTypeIs,
  iconIs, hasSome, hasAtLeast, taisenAbove, ...)
- ship-predicates.ts / equip-predicates.ts: shared named predicates,
  with predicates.ts as a barrel
- aaci/ moved as-is; aapb, oasw and sp-attack rewritten on the shared
  pool, dropping their duplicated local copies

aapb and sp-attack now take merged GameShip/GameEquip objects like
AACI/OASW already did; the callers merge in their selectors following
the existing oasw-indicator pattern. The old import paths
(views/utils/{aaci,oasw,aapb,sp_attack}) remain as deprecated shims so
external plugins keep working, with the aapb/sp_attack shims adapting
the historical pair-based signatures.

Predicates that only look shared are intentionally kept separate:
OASW's isFletcherClassOrKai (ctype-based) vs AACI's (id list), and
sp-attack's isBattleShip (includes stype 12) vs the shared isBattleship.

Behavior verified unchanged by new unit tests for aapb/oasw/sp-attack
(none existed before), including old-shim vs new-API equivalence, plus
a one-off differential run of the original implementations against the
refactored ones over 9000 randomized fleets/loadouts (identical output;
the only divergence is that sp-attack now returns false on short fleets
where the old code could throw).
