+++ views/utils/aaci.ts
@@ -1,0 +1,3 @@
+// Deprecated import path kept for plugin backward compatibility.
+// The module now lives in 'views/utils/combat/aaci'.
+export * from './combat/aaci'
