+++ views/utils/combat/aaci/table.ts
