+++ views/utils/combat/aaci/types.ts
@@ -1,0 +1,10 @@
+import type { EquipsPredicate, ShipPredicate } from '../types'
+
+export interface AACIEntry {
+  name?: string[]
+  id: number
+  fixed: number
+  modifier: number
+  shipValid: ShipPredicate
+  equipsValid: EquipsPredicate
+}
