+++ views/utils/combat/oasw.ts
@@ -1,0 +1,242 @@
+import type { GameEquip, GameShip } from './types'
+
+import { hasAtLeast, hasSome, shipIdIs, taisenAbove } from './combinators'
+import {
+  equipTais,
+  equipTaisAbove,
+  isASWAircraft,
+  isAutogyro,
+  isDepthCharge,
+  isDiveBomber,
+  isFixedWingASWAircraft,
+  isSeaplaneBomber,
+  isSonar,
+  isTorpedoBomber,
+} from './equip-predicates'
+import { isFubukiK3Go, isHyuuGaK2, isIsuzuK2, isKai, isTatsutaK2 } from './ship-predicates'
+
+type OASWPredicate = (ship: GameShip, equips: GameEquip[]) => boolean
+
+// Custom typed combinators (lodash's overSome/overEvery can't type mixed-arity predicates)
+const overSome =
+  (...fns: OASWPredicate[]): OASWPredicate =>
+  (ship, equips) =>
+    fns.some((f) => f(ship, equips))
+const overEvery =
+  (...fns: OASWPredicate[]): OASWPredicate =>
+  (ship, equips) =>
+    fns.every((f) => f(ship, equips))
+
+const isDE = (ship: GameShip) => ship.api_stype === 1
+const isCVL = (ship: GameShip) => ship.api_stype === 7
+
+const isJClassKai: OASWPredicate = (ship) =>
+  ship.api_ship_id === 394 || ship.api_ship_id === 893 || ship.api_ship_id === 906
+const isSamuelKai = shipIdIs(681)
+const isSamuelKaiNi = shipIdIs(920)
+const isFusoClassKaiNi: OASWPredicate = (ship) =>
+  ship.api_ship_id === 411 || ship.api_ship_id === 412
+const isFletcherClassKai = (ship: GameShip) => ship.api_ctype === 91 && isKai(ship)
+// NOTE: intentionally broader than the AACI isFletcherClassOrKai (ctype-based
+// so future Fletcher-class Kai are covered); keep them separate.
+const isFletcherClassOrKai: OASWPredicate = (ship) =>
+  ship.api_ship_id === 562 || ship.api_ship_id === 596 || isFletcherClassKai(ship)
+
+const isTaiyouClassKai: OASWPredicate = (ship) =>
+  ship.api_ship_id === 380 || ship.api_ship_id === 381
+const isTaiyouClassKaiNi: OASWPredicate = (ship) =>
+  ship.api_ship_id === 529 || ship.api_ship_id === 536
+const isMogamiClassKouKaiNi: OASWPredicate = (ship) =>
+  ship.api_ship_id === 508 || ship.api_ship_id === 509
+
+const isYuubariKaiNiTei = shipIdIs(624)
+
+const isKagaKaiNiGo = shipIdIs(646)
+
+const isShinShuMaruKai = shipIdIs(626)
+
+const isYamatoKaiNiJuu = shipIdIs(916)
+
+const isKumanomaru: OASWPredicate = (ship) => ship.api_ship_id === 943 || ship.api_ship_id === 948
+
+// focus on the 2nd argument of isOASW for func
+const overEquips =
+  (func: (equips: GameEquip[]) => boolean): OASWPredicate =>
+  (_ship: GameShip, equips: GameEquip[]) =>
+    func(equips)
+
+/*
+   - reference as of Jan 23, 2019: (TODO: not all implemented yet since Oct 18, 2018)
+
+       http://wikiwiki.jp/kancolle/?%C2%D0%C0%F8%C0%E8%C0%A9%C7%FA%CD%EB%B9%B6%B7%E2
+
+   - Shinyou-related OASW is kinda too messy at this point to be put here.
+
+ */
+// isOASW(ship: GameShip, equips: GameEquip[]): bool
+export const isOASW: OASWPredicate = overSome(
+  // 無条件に発動
+  isIsuzuK2,
+  isJClassKai,
+  isTatsutaK2,
+  isYuubariKaiNiTei,
+  isSamuelKai,
+  isFletcherClassOrKai,
+  isSamuelKaiNi,
+  isFubukiK3Go,
+  // 海防艦
+  overEvery(
+    isDE,
+    overSome(
+      // 必要対潜60 + ソナー
+      overEvery(taisenAbove(60), overEquips(hasSome(isSonar))),
+      // 必要対潜値75 + 装備のみの対潜値が合計4以上
+      overEvery(
+        taisenAbove(75),
+        overEquips((equips) => equips.map(equipTais).reduce((a, b) => a + b, 0) >= 4),
+      ),
+    ),
+  ),
+  overEvery(
+    (ship: GameShip) =>
+      [
+        // 駆逐
+        2,
+        // 軽巡
+        3,
+        // 雷巡
+        4,
+        // 練巡
+        21,
+        // 補給
+        22,
+      ].includes(ship.api_stype ?? -1),
+    taisenAbove(100),
+    overEquips(hasSome(isSonar)),
+  ),
+  // 大鷹型改 大鷹型改二 加賀改二護
+  overEvery(
+    overSome(isTaiyouClassKai, isTaiyouClassKaiNi, isKagaKaiNiGo),
+    overEquips(
+      hasSome(
+        (equip) =>
+          // 対潜値1以上の艦攻
+          (isTorpedoBomber(equip) && equipTaisAbove(1)(equip)) ||
+          // 対潜値1以上の艦爆
+          (isDiveBomber(equip) && equipTaisAbove(1)(equip)) ||
+          // 三式指揮連絡機(対潜) / カ号観測機
+          isASWAircraft(equip),
+      ),
+    ),
+  ),
+  // 護衛空母 (excluding 大鷹改 大鷹改二 最上型軽空母)
+  overEvery(
+    (s: GameShip) =>
+      !isTaiyouClassKai(s, []) &&
+      !isTaiyouClassKaiNi(s, []) &&
+      !isMogamiClassKouKaiNi(s, []) &&
+      isCVL(s),
+    overSome(
+      overEvery(
+        taisenAbove(65),
+        overEquips(
+          hasSome(
+            (equip) =>
+              // 対潜値7以上の艦攻
+              (isTorpedoBomber(equip) && equipTaisAbove(7)(equip)) ||
+              // 三式指揮連絡機(対潜) / カ号観測機
+              isASWAircraft(equip),
+          ),
+        ),
+      ),
+      overEvery(
+        taisenAbove(50),
+        overEquips(hasSome(isSonar)),
+        overEquips(
+          hasSome(
+            (equip) =>
+              // 対潜値7以上の艦攻
+              (isTorpedoBomber(equip) && equipTaisAbove(7)(equip)) ||
+              // 三式指揮連絡機(対潜) / カ号観測機
+              isASWAircraft(equip),
+          ),
+        ),
+      ),
+      overEvery(
+        taisenAbove(100),
+        overEquips(hasSome(isSonar)),
+        overEquips(
+          hasSome(
+            (equip) =>
+              // 対潜値1以上の艦攻
+              (isTorpedoBomber(equip) && equipTaisAbove(1)(equip)) ||
+              // 対潜値1以上の艦爆
+              (isDiveBomber(equip) && equipTaisAbove(1)(equip)),
+          ),
+        ),
+      ),
+    ),
+  ),
+  // 日向改二
+  overEvery(
+    isHyuuGaK2,
+    overSome(
+      // 対潜値12以上のオートジャイロ
+      overEquips(hasSome((equip) => isAutogyro(equip) && equipTaisAbove(12)(equip))),
+      // オートジャイロ二機
+      overEquips(hasAtLeast(isAutogyro, 2)),
+    ),
+  ),
+  // 神州丸改 大和改二重
+  overEvery(
+    overSome(isShinShuMaruKai, isYamatoKaiNiJuu),
+    taisenAbove(100),
+    overEquips(
+      hasSome(
+        (equip) =>
+          // 回転翼機
+          isAutogyro(equip) ||
+          // 水上爆撃機
+          isSeaplaneBomber(equip),
+      ),
+    ),
+    overEquips(hasSome(isSonar)),
+  ),
+  // 熊野丸/改
+  overEvery(
+    isKumanomaru,
+    taisenAbove(100),
+    overEquips(hasSome(isSonar)),
+    overEquips(
+      hasSome(
+        (equip) =>
+          // 対潜値1以上の艦爆
+          (isDiveBomber(equip) && equipTaisAbove(1)(equip)) ||
+          // オートジャイロ機
+          isAutogyro(equip) ||
+          // 対潜哨戒機
+          isFixedWingASWAircraft(equip),
+      ),
+    ),
+  ),
+  // 扶桑改二 山城改二
+  overEvery(
+    isFusoClassKaiNi,
+    taisenAbove(100),
+    overEquips(hasSome(isSonar)),
+    overEquips(
+      hasSome(
+        (equip) =>
+          // 水上爆撃機
+          isSeaplaneBomber(equip) ||
+          // オートジャイロ機
+          isAutogyro(equip) ||
+          // 爆雷投射機/爆雷
+          isDepthCharge(equip),
+      ),
+    ),
+  ),
+)
+
+// Backward compatibility
+export const isOASWWith = () => isOASW
