+++ views/utils/combat/predicates.ts
@@ -1,0 +1,4 @@
+// Barrel for the shared predicate pool used by the combat mechanics modules.
+export * from './combinators'
+export * from './ship-predicates'
+export * from './equip-predicates'
